+++ crates/rspack_diff/src/dist_diff/mod.rs
@@ -1,1 +1,4 @@
-pub mod command;
+mod options;
+mod command;
+pub use options::*;
+pub use command::*;
