+++ crates/rspack_diff/src/result.rs
@@ -1,4 +1,4 @@
-use crate::dist_diff::command::DistDiffResult;
+use crate::dist_diff::DistDiffResult;
 #[derive(Debug)]
 pub enum CliRunResult {
   DistDiffResult(DistDiffResult),
