+++ crates/rspack_diff/src/runner.rs
@@ -9,6 +9,10 @@ pub trait Runner {
   const NAME: &'static str;
   const ABOUT: &'static str;
   type Options: RunnerOptions;
+  fn command() -> Command {
+    let cmd = Command::new(Self::NAME).about(Self::ABOUT);
+    Self::Options::build_args(cmd)
+  }
   fn run(&self) -> result::CliRunResult;
   fn new(options: Self::Options) -> Self;
 }
